complete initial project card
